+++ resources/views/livewire/login-shortcut.blade.php
@@ -2,12 +2,12 @@
     @if ($available)
         <x-filament::section>
             <x-slot name="heading">
-                {{ __('filament-developer-login::messages.heading') }}
+                {{ __('filament-login-shortcut::messages.heading') }}
             </x-slot>
 
             @if ($nonLocal)
                 <x-filament::callout color="warning">
-                    {{ __('filament-developer-login::messages.warning', ['environments' => $environments]) }}
+                    {{ __('filament-login-shortcut::messages.warning', ['environments' => $environments]) }}
                 </x-filament::callout>
             @endif
 
@@ -15,7 +15,7 @@
                 {{ $this->form }}
 
                 <x-filament::button type="button" wire:click="login" wire:loading.attr="disabled" :disabled="blank($data['selectedIdentifier'] ?? null)" style="width: 100%;">
-                    {{ __('filament-developer-login::messages.login_button') }}
+                    {{ __('filament-login-shortcut::messages.login_button') }}
                 </x-filament::button>
             </div>
         </x-filament::section>